+++ Mailer/migrations/20210414065725_missing_mail_sender_id_index.php
@@ -1,0 +1,14 @@
+<?php
+declare(strict_types=1);
+
+use Phinx\Migration\AbstractMigration;
+
+final class MissingMailSenderIdIndex extends AbstractMigration
+{
+    public function change(): void
+    {
+        $this->table('mail_logs')
+            ->addIndex('mail_sender_id')
+            ->update();
+    }
+}